Build: Fix precios y details

+++ src/app/admin/extras/create/page.jsx
@@ -81,7 +81,7 @@ export default function ExtraCreateAdminPage() {
                         classNames={{
                             base: `${styles.input}`,
                         }}
-                        onValueChange={(value) => setInputPrice(parseInt(value))}
+                        onValueChange={(value) => setInputPrice(parseFloat(value))}
                     />
 
                 </div>
